refactor to use only 1 module
