+++ requests.http
@@ -1,9 +1,9 @@
 @host = http://localhost:3203
-@token = eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InVzZXIxQGRvbWFpbi5jb20iLCJpYXQiOjE3MDg4MDI1NTEsImV4cCI6MTcwODgwNjE1MX0.W2brgINPz4XYim9E4GikvKZ9JbQNXFf0vx2au2oAYh8
+@token = eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InVzZXIxQGRvbWFpbi5jb20iLCJpYXQiOjE3MTAxNzA1NTEsImV4cCI6MTcxMDE3NDE1MX0.rUa3xwDnoOgRPuj3t3l9t5x-CoA3ujjwHWiHoAitMSk
 
 
 ### sign in user
-POST {{host}}/auth/sign-in
+POST {{host}}/users/sign-in
 Content-Type: application/json
 
 {
@@ -12,7 +12,7 @@ Content-Type: application/json
 }
 
 ### sign up user
-POST {{host}}/auth/sign-up
+POST {{host}}/users/sign-up
 Content-Type: application/json
 
 {
@@ -23,7 +23,7 @@ Content-Type: application/json
 
 
 ### sing out user
-POST {{host}}/auth/sign-out
+POST {{host}}/users/sign-out
 Content-Type: application/json
 Authorization: Bearer {{token}}
