+++ src/resources/user-auth/dtos/user-dtos/responses.dto.ts
@@ -1,5 +1,5 @@
 import { OmitType } from '@nestjs/swagger';
-import { UserModel } from '../models/user.model';
+import { UserModel } from '../../models/user/user.model';
 
 export class UserResponse extends OmitType(UserModel, [
   'hashedPassword',
