+++ src/resources/user-auth/exceptions/user-custom-exceptions.ts
