+++ src/resources/user-auth/models/auth/auth-info.model.ts
